Move function to common header
