+++ dev/Common/Common.vcxitems
@@ -23,6 +23,7 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)AppModel.Package.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)AppModel.PackageGraph.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Microsoft.Foundation.String.h" />
+    <ClInclude Include="$(MSBuildThisFileDirectory)Microsoft.RoApi.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Microsoft.Utf8.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)NotificationTelemetryHelper.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Security.IntegrityLevel.h" />
