+++ dev/Common/Common.vcxitems.filters
@@ -38,6 +38,9 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)NotificationTelemetryHelper.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="$(MSBuildThisFileDirectory)Microsoft.RoApi.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="$(MSBuildThisFileDirectory)WindowsAppRuntime.SelfContained.cpp">
